fixed two compiler warnings in vhost code

+++ src/ircd.c
@@ -966,7 +966,7 @@ normal user.\n");
       if ((aconf->passwd[0] != '\0') && (aconf->passwd[0] != '*'))
 	vaddr = inet_addr(aconf->passwd);
       else
-        vaddr = NULL;
+        vaddr = (u_long) NULL;
 
       if (inetport(&me, star, portnum, vaddr))
       {
